voice_agents/personal_assistant/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
av==15.1.0
backoff==2.2.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.0
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-genai==1.50.1
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.7
langchain-core==1.0.5
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
livekit==1.0.19
livekit-agents==1.2.18
livekit-api==1.0.7
livekit-blingfire==1.0.0
livekit-plugins-elevenlabs==1.2.18
livekit-plugins-google==1.2.18
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.2.18
livekit-plugins-silero==1.2.18
livekit-plugins-turn-detector==1.2.18
livekit-protocol==1.1.0
MarkupSafe==3.0.3
mcp==1.21.1
mem0ai==1.0.1
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.3.4
onnxruntime==1.23.1
openai==2.8.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pillow==12.0.0
portalocker==3.2.0
posthog==7.0.1
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
types-protobuf==6.32.1.20251105
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
